+++ src/Projection/InMemoryEventStoreProjection.php
@@ -12,28 +12,132 @@ declare(strict_types=1);
 
 namespace Prooph\EventStore\Projection;
 
-use Prooph\EventStore\Exception\RuntimeException;
+use ArrayIterator;
+use Closure;
+use Iterator;
+use Prooph\Common\Messaging\Message;
+use Prooph\EventStore\Exception;
 use Prooph\EventStore\InMemoryEventStore;
+use Prooph\EventStore\Stream;
+use Prooph\EventStore\StreamName;
+use Prooph\EventStore\Util\ArrayCache;
 
-final class InMemoryEventStoreProjection extends AbstractProjection
+final class InMemoryEventStoreProjection implements Projection
 {
     /**
      * @var array
      */
     private $knownStreams;
 
-    protected function buildKnownStreams()
+    /**
+     * @var string
+     */
+    private $name;
+
+    /**
+     * @var ArrayCache
+     */
+    private $cachedStreamNames;
+
+    /**
+     * @var InMemoryEventStore
+     */
+    private $eventStore;
+
+    /**
+     * @var array
+     */
+    private $streamPositions;
+
+    /**
+     * @var array
+     */
+    private $state = [];
+
+    /**
+     * @var callable|null
+     */
+    private $initCallback;
+
+    /**
+     * @var Closure|null
+     */
+    private $handler;
+
+    /**
+     * @var array
+     */
+    private $handlers = [];
+
+    /**
+     * @var boolean
+     */
+    private $isStopped = false;
+
+    /**
+     * @var ?string
+     */
+    private $currentStreamName = null;
+
+    public function __construct(InMemoryEventStore $eventStore, string $name, int $cacheSize)
     {
+        $this->eventStore = $eventStore;
+        $this->name = $name;
+        $this->cachedStreamNames = new ArrayCache($cacheSize);
+
         $reflectionProperty = new \ReflectionProperty(get_class($this->eventStore), 'streams');
         $reflectionProperty->setAccessible(true);
 
         $this->knownStreams = array_keys($reflectionProperty->getValue($this->eventStore));
     }
 
+    public function init(Closure $callback): Query
+    {
+        if (null !== $this->initCallback) {
+            throw new Exception\RuntimeException('Projection already initialized');
+        }
+
+        $callback = Closure::bind($callback, $this->createHandlerContext($this->currentStreamName));
+
+        $result = $callback();
+
+        if (is_array($result)) {
+            $this->state = $result;
+        }
+
+        $this->initCallback = $callback;
+
+        return $this;
+    }
+
+    public function fromStream(string $streamName): Query
+    {
+        if (null !== $this->streamPositions) {
+            throw new Exception\RuntimeException('From was already called');
+        }
+
+        $this->streamPositions = [$streamName => 0];
+
+        return $this;
+    }
+
+    public function fromStreams(string ...$streamNames): Query
+    {
+        if (null !== $this->streamPositions) {
+            throw new Exception\RuntimeException('From was already called');
+        }
+
+        foreach ($streamNames as $streamName) {
+            $this->streamPositions[$streamName] = 0;
+        }
+
+        return $this;
+    }
+
     public function fromCategory(string $name): Query
     {
         if (null !== $this->streamPositions) {
-            throw new RuntimeException('from was already called');
+            throw new Exception\RuntimeException('from was already called');
         }
 
         $this->streamPositions = [];
@@ -50,7 +154,7 @@ final class InMemoryEventStoreProjection extends AbstractProjection
     public function fromCategories(string ...$names): Query
     {
         if (null !== $this->streamPositions) {
-            throw new RuntimeException('from was already called');
+            throw new Exception\RuntimeException('from was already called');
         }
 
         $this->streamPositions = [];
@@ -70,7 +174,7 @@ final class InMemoryEventStoreProjection extends AbstractProjection
     public function fromAll(): Query
     {
         if (null !== $this->streamPositions) {
-            throw new RuntimeException('from was already called');
+            throw new Exception\RuntimeException('from was already called');
         }
 
         $this->streamPositions = [];
@@ -86,27 +190,231 @@ final class InMemoryEventStoreProjection extends AbstractProjection
         return $this;
     }
 
-    public function __construct(InMemoryEventStore $eventStore, string $name, int $cacheSize, int $persistBlockSize)
+    public function when(array $handlers): Query
     {
-        parent::__construct($eventStore, $name, $cacheSize, $persistBlockSize);
+        if (null !== $this->handler || ! empty($this->handlers)) {
+            throw new Exception\RuntimeException('When was already called');
+        }
 
-        $this->buildKnownStreams();
+        foreach ($handlers as $eventName => $handler) {
+            if (! is_string($eventName)) {
+                throw new Exception\InvalidArgumentException('Invalid event name given, string expected');
+            }
+
+            if (! $handler instanceof Closure) {
+                throw new Exception\InvalidArgumentException('Invalid handler given, Closure expected');
+            }
+
+            $this->handlers[$eventName] = Closure::bind($handler, $this->createHandlerContext($this->currentStreamName));
+        }
+
+        return $this;
+    }
+
+    public function whenAny(Closure $handler): Query
+    {
+        if (null !== $this->handler || ! empty($this->handlers)) {
+            throw new Exception\RuntimeException('When was already called');
+        }
+
+        $this->handler = Closure::bind($handler, $this->createHandlerContext($this->currentStreamName));
+
+        return $this;
+    }
+
+    public function stop(): void
+    {
+        $this->isStopped = true;
+    }
+
+    public function getState(): array
+    {
+        return $this->state;
+    }
+
+    public function getName(): string
+    {
+        return $this->name;
+    }
+
+    public function emit(Message $event): void
+    {
+        $this->linkTo($this->name, $event);
+    }
+
+    public function linkTo(string $streamName, Message $event): void
+    {
+        $sn = new StreamName($streamName);
+
+        if ($this->cachedStreamNames->has($streamName)) {
+            $append = true;
+        } else {
+            $this->cachedStreamNames->rollingAppend($streamName);
+            $append = $this->eventStore->hasStream($sn);
+        }
+
+        if ($append) {
+            $this->eventStore->appendTo($sn, new ArrayIterator([$event]));
+        } else {
+            $this->eventStore->create(new Stream($sn, new ArrayIterator([$event])));
+        }
+    }
+
+    public function reset(): void
+    {
+        if (null !== $this->streamPositions) {
+            $this->streamPositions = array_map(
+                function (): int {
+                    return 0;
+                },
+                $this->streamPositions
+            );
+        }
+
+        $callback = $this->initCallback;
+
+        if (is_callable($callback)) {
+            $result = $callback();
+
+            if (is_array($result)) {
+                $this->state = $result;
+
+                return;
+            }
+        }
+
+        $this->state = [];
+
+        $this->eventStore->delete(new StreamName($this->name));
+    }
+
+    public function run(bool $keepRunning = true): void
+    {
+        if (null === $this->streamPositions
+            || (null === $this->handler && empty($this->handlers))
+        ) {
+            throw new Exception\RuntimeException('No handlers configured');
+        }
+
+        do {
+            if (! $this->eventStore->hasStream(new StreamName($this->name))) {
+                $this->eventStore->create(new Stream(new StreamName($this->name), new ArrayIterator()));
+            }
+
+            $singleHandler = null !== $this->handler;
+
+            foreach ($this->streamPositions as $streamName => $position) {
+                try {
+                    $stream = $this->eventStore->load(new StreamName($streamName), $position + 1);
+                } catch (Exception\StreamNotFound $e) {
+                    // no newer events found
+                    continue;
+                }
+
+                if ($singleHandler) {
+                    $this->handleStreamWithSingleHandler($streamName, $stream->streamEvents());
+                } else {
+                    $this->handleStreamWithHandlers($streamName, $stream->streamEvents());
+                }
+
+                if ($this->isStopped) {
+                    break;
+                }
+            }
+        } while ($keepRunning && ! $this->isStopped);
     }
 
     public function delete(bool $deleteEmittedEvents): void
     {
         if ($deleteEmittedEvents) {
-            $this->resetProjection();
+            $this->eventStore->delete(new StreamName($this->name));
         }
     }
 
-    protected function load(): void
+    private function handleStreamWithSingleHandler(string $streamName, Iterator $events): void
     {
-        // InMemoryEventStoreProjection cannot load
+        $this->currentStreamName = $streamName;
+        $handler = $this->handler;
+
+        foreach ($events as $event) {
+            /* @var Message $event */
+            $this->streamPositions[$streamName]++;
+
+            $result = $handler($this->state, $event);
+
+            if (is_array($result)) {
+                $this->state = $result;
+            }
+
+            if ($this->isStopped) {
+                break;
+            }
+        }
     }
 
-    protected function persist(): void
+    private function handleStreamWithHandlers(string $streamName, Iterator $events): void
     {
-        // InMemoryEventStoreProjection cannot persist
+        $this->currentStreamName = $streamName;
+
+        foreach ($events as $event) {
+            /* @var Message $event */
+            $this->streamPositions[$streamName]++;
+
+            if (! isset($this->handlers[$event->messageName()])) {
+                continue;
+            }
+
+            $handler = $this->handlers[$event->messageName()];
+            $result = $handler($this->state, $event);
+
+            if (is_array($result)) {
+                $this->state = $result;
+            }
+
+            if ($this->isStopped) {
+                break;
+            }
+        }
+    }
+
+    private function createHandlerContext(?string &$streamName)
+    {
+        return new class($this, $streamName) {
+            /**
+             * @var Projection
+             */
+            private $projection;
+
+            /**
+             * @var ?string
+             */
+            private $streamName;
+
+            public function __construct(Projection $projection, ?string &$streamName)
+            {
+                $this->projection = $projection;
+                $this->streamName = &$streamName;
+            }
+
+            public function stop(): void
+            {
+                $this->projection->stop();
+            }
+
+            public function linkTo(string $streamName, Message $event): void
+            {
+                $this->projection->linkTo($streamName, $event);
+            }
+
+            public function emit(Message $event): void
+            {
+                $this->projection->emit($event);
+            }
+
+            public function streamName(): ?string
+            {
+                return $this->streamName;
+            }
+        };
     }
 }
