+++ src/Projection/InMemoryEventStoreQuery.php
@@ -12,35 +12,117 @@ declare(strict_types=1);
 
 namespace Prooph\EventStore\Projection;
 
-use Prooph\EventStore\Exception\RuntimeException;
+use Closure;
+use Iterator;
+use Prooph\Common\Messaging\Message;
+use Prooph\EventStore\Exception;
 use Prooph\EventStore\InMemoryEventStore;
+use Prooph\EventStore\StreamName;
 
-final class InMemoryEventStoreQuery extends AbstractQuery
+final class InMemoryEventStoreQuery implements Query
 {
     /**
      * @var array
      */
     private $knownStreams;
 
+    /**
+     * @var InMemoryEventStore
+     */
+    private $eventStore;
+
+    /**
+     * @var array
+     */
+    private $streamPositions;
+
+    /**
+     * @var array
+     */
+    private $state = [];
+
+    /**
+     * @var callable|null
+     */
+    private $initCallback;
+
+    /**
+     * @var Closure|null
+     */
+    private $handler;
+
+    /**
+     * @var array
+     */
+    private $handlers = [];
+
+    /**
+     * @var boolean
+     */
+    private $isStopped = false;
+
+    /**
+     * @var ?string
+     */
+    private $currentStreamName = null;
+
     public function __construct(InMemoryEventStore $eventStore)
     {
-        parent::__construct($eventStore);
+        $this->eventStore = $eventStore;
 
-        $this->buildKnownStreams();
-    }
-
-    protected function buildKnownStreams()
-    {
         $reflectionProperty = new \ReflectionProperty(get_class($this->eventStore), 'streams');
         $reflectionProperty->setAccessible(true);
 
         $this->knownStreams = array_keys($reflectionProperty->getValue($this->eventStore));
     }
 
+    public function init(Closure $callback): Query
+    {
+        if (null !== $this->initCallback) {
+            throw new Exception\RuntimeException('Projection already initialized');
+        }
+
+        $callback = Closure::bind($callback, $this->createHandlerContext($this->currentStreamName));
+
+        $result = $callback();
+
+        if (is_array($result)) {
+            $this->state = $result;
+        }
+
+        $this->initCallback = $callback;
+
+        return $this;
+    }
+
+    public function fromStream(string $streamName): Query
+    {
+        if (null !== $this->streamPositions) {
+            throw new Exception\RuntimeException('From was already called');
+        }
+
+        $this->streamPositions = [$streamName => 0];
+
+        return $this;
+    }
+
+    public function fromStreams(string ...$streamNames): Query
+    {
+        if (null !== $this->streamPositions) {
+            throw new Exception\RuntimeException('From was already called');
+        }
+
+        foreach ($streamNames as $streamName) {
+            $this->streamPositions[$streamName] = 0;
+        }
+
+        return $this;
+    }
+
     public function fromCategory(string $name): Query
     {
         if (null !== $this->streamPositions) {
-            throw new RuntimeException('from was already called');
+            throw new Exception\RuntimeException('from was already called');
         }
 
         $this->streamPositions = [];
@@ -57,7 +139,7 @@ final class InMemoryEventStoreQuery extends AbstractQuery
     public function fromCategories(string ...$names): Query
     {
         if (null !== $this->streamPositions) {
-            throw new RuntimeException('from was already called');
+            throw new Exception\RuntimeException('from was already called');
         }
 
         $this->streamPositions = [];
@@ -77,7 +159,7 @@ final class InMemoryEventStoreQuery extends AbstractQuery
     public function fromAll(): Query
     {
         if (null !== $this->streamPositions) {
-            throw new RuntimeException('from was already called');
+            throw new Exception\RuntimeException('from was already called');
         }
 
         $this->streamPositions = [];
@@ -91,5 +173,174 @@ final class InMemoryEventStoreQuery extends AbstractQuery
         }
 
         return $this;
+    }
+
+    public function when(array $handlers): Query
+    {
+        if (null !== $this->handler || ! empty($this->handlers)) {
+            throw new Exception\RuntimeException('When was already called');
+        }
+
+        foreach ($handlers as $eventName => $handler) {
+            if (! is_string($eventName)) {
+                throw new Exception\InvalidArgumentException('Invalid event name given, string expected');
+            }
+
+            if (! $handler instanceof Closure) {
+                throw new Exception\InvalidArgumentException('Invalid handler given, Closure expected');
+            }
+
+            $this->handlers[$eventName] = Closure::bind($handler, $this->createHandlerContext($this->currentStreamName));
+        }
+
+        return $this;
+    }
+
+    public function whenAny(Closure $handler): Query
+    {
+        if (null !== $this->handler || ! empty($this->handlers)) {
+            throw new Exception\RuntimeException('When was already called');
+        }
+
+        $this->handler = Closure::bind($handler, $this->createHandlerContext($this->currentStreamName));
+
+        return $this;
+    }
+
+    public function reset(): void
+    {
+        if (null !== $this->streamPositions) {
+            $this->streamPositions = array_map(
+                function (): int {
+                    return 0;
+                },
+                $this->streamPositions
+            );
+        }
+
+        $callback = $this->initCallback;
+
+        if (is_callable($callback)) {
+            $result = $callback();
+
+            if (is_array($result)) {
+                $this->state = $result;
+
+                return;
+            }
+        }
+
+        $this->state = [];
+    }
+
+    public function run(): void
+    {
+        if (null === $this->streamPositions
+            || (null === $this->handler && empty($this->handlers))
+        ) {
+            throw new Exception\RuntimeException('No handlers configured');
+        }
+
+        $singleHandler = null !== $this->handler;
+
+        foreach ($this->streamPositions as $streamName => $position) {
+            $stream = $this->eventStore->load(new StreamName($streamName), $position + 1);
+
+            if ($singleHandler) {
+                $this->handleStreamWithSingleHandler($streamName, $stream->streamEvents());
+            } else {
+                $this->handleStreamWithHandlers($streamName, $stream->streamEvents());
+            }
+
+            if ($this->isStopped) {
+                break;
+            }
+        }
+    }
+
+    public function stop(): void
+    {
+        $this->isStopped = true;
+    }
+
+    public function getState(): array
+    {
+        return $this->state;
+    }
+
+    private function handleStreamWithSingleHandler(string $streamName, Iterator $events): void
+    {
+        $this->currentStreamName = $streamName;
+        $handler = $this->handler;
+
+        foreach ($events as $event) {
+            /* @var Message $event */
+            $this->streamPositions[$streamName]++;
+
+            $result = $handler($this->state, $event);
+
+            if (is_array($result)) {
+                $this->state = $result;
+            }
+
+            if ($this->isStopped) {
+                break;
+            }
+        }
+    }
+
+    private function handleStreamWithHandlers(string $streamName, Iterator $events): void
+    {
+        $this->currentStreamName = $streamName;
+        foreach ($events as $event) {
+            /* @var Message $event */
+            $this->streamPositions[$streamName]++;
+
+            if (! isset($this->handlers[$event->messageName()])) {
+                continue;
+            }
+
+            $handler = $this->handlers[$event->messageName()];
+            $result = $handler($this->state, $event);
+
+            if (is_array($result)) {
+                $this->state = $result;
+            }
+
+            if ($this->isStopped) {
+                break;
+            }
+        }
+    }
+
+    private function createHandlerContext(?string &$streamName)
+    {
+        return new class($this, $streamName) {
+            /**
+             * @var Query
+             */
+            private $query;
+
+            /**
+             * @var ?string
+             */
+            private $streamName;
+
+            public function __construct(Query $query, ?string &$streamName)
+            {
+                $this->query = $query;
+                $this->streamName = &$streamName;
+            }
+
+            public function stop(): void
+            {
+                $this->query->stop();
+            }
+
+            public function streamName(): ?string
+            {
+                return $this->streamName;
+            }
+        };
     }
 }
