+++ tests/Mock/ReadModelMock.php
@@ -40,7 +40,7 @@ class ReadModelMock extends AbstractReadModel
 
     public function hasKey(string $key): bool
     {
-        return array_key_exists($key, $this->storage);
+        return is_array($this->storage) && array_key_exists($key, $this->storage);
     }
 
     public function read(string $key)
